tahir handover
