+++ latch_metadata/__init__.py
@@ -1,17 +1,9 @@
-import sys
-
-import typing_extensions
-from flytekit.core.annotation import FlyteAnnotation
-from latch.ldata.path import LPath
-from latch.types import metadata
-from latch.types.directory import LatchDir, LatchOutputDir
-from latch.types.file import LatchFile
+from latch.types.directory import LatchDir
 from latch.types.metadata import (
     Fork,
     ForkBranch,
     LatchAuthor,
     NextflowMetadata,
-    NextflowParameter,
     NextflowRuntimeResources,
     Params,
     Section,
@@ -22,7 +14,7 @@ from latch.types.metadata import (
 from .parameters import generated_parameters
 
 flow = [
-    Section("Input/Output", Params("input", "outdir", "run_name")),
+    Section("Inputs", Params("input")),
     Section(
         "Reference Genome Options",
         Fork(
@@ -87,6 +79,12 @@ flow = [
             ),
         ),
     ),
+    Section(
+        "Output Directory",
+        Params("run_name"),
+        Text("Parent directory for outputs"),
+        Params("outdir"),
+    ),
     Spoiler(
         "Optional Parameters",
         Spoiler(
@@ -128,6 +126,8 @@ flow = [
         ),
     ),
 ]
+
+
 NextflowMetadata(
     display_name="nf-core/atacseq",
     author=LatchAuthor(
